+++ components/DungeonIcon.tsx
